Translate OTEL log's InstrumentationScope as logger name
InstrumentationScope usually contains the name of the instrumentation
library. However, in the case of logs, it may also contain the name of the
logger if the log source defines it [1]. An instrumentation library usually
has a version set, so assume that the name without the version is the name
of the logger.

[1] https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/logs/api.md#get-a-logger

Signed-off-by: Oldřich Jedlička <oldium.pro@gmail.com>